First Successful Build with MPLAB X IDE 3.35
usart.h does not need to be specified anymore., so it was moved to only
be used by other compliers.  Notes added for installing the peripheral
library, which is no longer included with XC8.  They are phasing
everything to MPlab Code Configurator, however these processors are not
supported by it, so we must use the peripheral library.